Add a batch tune mode that sweeps every knob in a stage or sub-block
Add run_scan_batch and a 'tune scan-all' verb (plus Pipeline / api
wrappers) that sweep every knob matched by a path-prefix selector on its
default grid, so a whole stage or sub-block can be reviewed in one pass
instead of knob-by-knob. A knob whose required stage is absent is
captured as a failed BatchItem and the batch continues. Take the
magnitude of the complex spectrum directly in the noise overlay rather
than casting to real first, which dropped the imaginary part.

+++ src/ftmwpipeline/_internal/tuning/plots.py
@@ -445,7 +445,7 @@ def plot_noise_sweep(spec: Any, rows: List[Any], ctx: Any) -> Any:
 
         ft = ftmw.compute_ft(ctx.ftmw_path)
         f_ghz = np.asarray(ft.freq_array, dtype=float) / 1000.0
-        mag = np.abs(np.asarray(ft.complex_spectrum, dtype=float))
+        mag = np.abs(np.asarray(ft.complex_spectrum))
         ax3.plot(f_ghz, mag, lw=0.4, color="0.6", label="|FT|", zorder=1)
         sigma_max = 0.0
         for row, color in zip(rows, colors):
